Improves active read of values.
Now it is possible to choose how values for metrics should be read
actively. The default method is as already known sending a GroupRead
message to the GroupAddress of the value. New is to send a GroupWrite
message to a different GroupAddress which acts as a trigger for the
device to send the requested values.